requirements.txt
pdfplumber
transformers
sentence-transformers
python-pptx
Pillow
moviepy
pyttsx3
torch
tqdm
nltk